+++ packages/core/client/src/schema-component/antd/calendar/Nav.tsx
@@ -2,7 +2,7 @@ import { LeftOutlined, RightOutlined } from '@ant-design/icons';
 import { observer } from '@formily/react';
 import { Button } from 'antd';
 import React, { useContext } from 'react';
-import { navigate } from 'react-big-calendar/lib/utils/constants';
+import { Navigate } from 'react-big-calendar/dist/react-big-calendar.esm';
 import { useDesignable } from '../../hooks';
 import { CalendarToolbarContext } from './context';
 
@@ -12,8 +12,8 @@ export const Nav = observer(
     const { onNavigate } = useContext(CalendarToolbarContext);
     return (
       <div className="ant-btn-group">
-        <Button icon={<LeftOutlined />} onClick={() => onNavigate(navigate.PREVIOUS)}></Button>
-        <Button icon={<RightOutlined />} onClick={() => onNavigate(navigate.NEXT)}></Button>
+        <Button icon={<LeftOutlined />} onClick={() => onNavigate(Navigate.PREVIOUS)}></Button>
+        <Button icon={<RightOutlined />} onClick={() => onNavigate(Navigate.NEXT)}></Button>
         <DesignableBar />
       </div>
     );
