+++ packages/core/client/src/schema-component/antd/calendar/Today.tsx
@@ -1,7 +1,7 @@
 import { observer } from '@formily/react';
 import { Button } from 'antd';
 import React, { useContext } from 'react';
-import { navigate } from 'react-big-calendar/lib/utils/constants';
+import { Navigate } from 'react-big-calendar/dist/react-big-calendar.esm';
 import { useTranslation } from 'react-i18next';
 import { useDesignable } from '../../hooks';
 import { CalendarToolbarContext } from './context';
@@ -14,7 +14,7 @@ export const Today = observer(
     return (
       <Button
         onClick={() => {
-          onNavigate(navigate.TODAY);
+          onNavigate(Navigate.TODAY);
         }}
       >
         {t('Today')}
