+++ packages/core/client/src/schema-component/antd/date-picker/DatePicker.tsx
@@ -3,7 +3,7 @@ import { DatePicker as AntdDatePicker } from 'antd';
 import type {
   DatePickerProps as AntdDatePickerProps,
   RangePickerProps as AntdRangePickerProps,
-} from 'antd/lib/date-picker';
+} from 'antd/es/date-picker';
 import React from 'react';
 import { useTranslation } from 'react-i18next';
 import { ReadPretty } from './ReadPretty';
