+++ packages/core/client/src/schema-component/antd/date-picker/ReadPretty.tsx
@@ -1,10 +1,10 @@
-import { usePrefixCls } from '@formily/antd/lib/__builtins__';
+import { usePrefixCls } from '@formily/antd/esm/__builtins__';
 import { isArr } from '@formily/shared';
 import { getDefaultFormat, str2moment } from '@nocobase/utils/client';
 import type {
   DatePickerProps as AntdDatePickerProps,
   RangePickerProps as AntdRangePickerProps,
-} from 'antd/lib/date-picker';
+} from 'antd/es/date-picker';
 import cls from 'classnames';
 import moment from 'moment';
 import React from 'react';
