+++ packages/core/client/src/schema-component/antd/input-number/ReadPretty.tsx
@@ -1,7 +1,7 @@
 import { isValid } from '@formily/shared';
 import { toFixedByStep } from '@nocobase/utils/client';
-import type { InputProps } from 'antd/lib/input';
-import type { InputNumberProps } from 'antd/lib/input-number';
+import type { InputProps } from 'antd/es/input';
+import type { InputNumberProps } from 'antd/es/input-number';
 import React from 'react';
 
 export const ReadPretty: React.FC<InputProps & InputNumberProps> = (props: any) => {
