+++ packages/core/client/src/schema-component/antd/input/ReadPretty.tsx
@@ -1,12 +1,12 @@
-import { usePrefixCls } from '@formily/antd/lib/__builtins__';
-import { InputProps, TextAreaProps } from 'antd/lib/input';
+import { css, cx } from '@emotion/css';
+import { usePrefixCls } from '@formily/antd/esm/__builtins__';
+import { Typography } from 'antd';
+import { InputProps, TextAreaProps } from 'antd/es/input';
 import cls from 'classnames';
 import React from 'react';
 import { useCompile } from '../..';
 import { EllipsisWithTooltip } from './EllipsisWithTooltip';
 import { HTMLEncode } from './shared';
-import { cx, css } from '@emotion/css';
-import { Typography } from 'antd';
 
 type Composed = {
   Input: React.FC<InputProps & { ellipsis?: any }>;
