+++ packages/core/client/src/schema-component/antd/time-picker/TimePicker.tsx
@@ -1,7 +1,7 @@
-import { formatMomentValue, momentable } from '@formily/antd/lib/__builtins__';
+import { formatMomentValue, momentable } from '@formily/antd/esm/__builtins__';
 import { connect, mapProps, mapReadPretty } from '@formily/react';
 import { TimePicker as AntdTimePicker } from 'antd';
-import { TimePickerProps as AntdTimePickerProps, TimeRangePickerProps } from 'antd/lib/time-picker';
+import { TimePickerProps as AntdTimePickerProps, TimeRangePickerProps } from 'antd/es/time-picker';
 import moment from 'moment';
 import { ReadPretty } from './ReadPretty';
