+++ packages/core/dumi-theme-nocobase/src/builtins/Tree.tsx
@@ -1,9 +1,9 @@
-import type { ReactNode, ComponentProps } from 'react';
-import React, { useEffect, useState, createRef } from 'react';
+import type { CSSMotionProps, MotionEndEventHandler, MotionEventHandler } from 'rc-motion';
 import type { TreeProps } from 'rc-tree';
 import Tree from 'rc-tree';
-import type { EventDataNode } from 'rc-tree/lib/interface';
-import type { CSSMotionProps, MotionEventHandler, MotionEndEventHandler } from 'rc-motion';
+import type { EventDataNode } from 'rc-tree/es/interface';
+import type { ComponentProps, ReactNode } from 'react';
+import React, { createRef, useEffect, useState } from 'react';
 import './Tree.less';
 
 const FileOutlined = <svg xmlns="http://www.w3.org/2000/svg" className="__dumi-site-tree-icon icon-file" fill="currentcolor" viewBox="0 0 1024 1024"><path d="M854.6 288.6L639.4 73.4c-6-6-14.1-9.4-22.6-9.4H192c-17.7 0-32 14.3-32 32v832c0 17.7 14.3 32 32 32h640c17.7 0 32-14.3 32-32V311.3c0-8.5-3.4-16.7-9.4-22.7zM790.2 326H602V137.8L790.2 326zm1.8 562H232V136h302v216a42 42 0 0 0 42 42h216v494z"/></svg>;
